Reword the introduction "Example" section
A few small changes to the example in the intro page. Other than the wording, the main change is the removal of the footer about how websites can load resources from each other -- I feel like this is prominently covered in the first two sections of the page and directly in text in this section, so the footnote felt a bit superfluous.