RIP is required for client-side inflation error side-channel mitigation
It is possible to detect redirections using fetch using the same technique of client-side inflation errors.

For example, on Chrome:
`await fetch(`https://bank.com/redir#<BIG-fragment>`, {method: 'GET', mode :'no-cors', credentials: 'include'});`

The server redirects to `/bigger` based on a cookie which is included in the fetch request. The first request succeeds, but the second request (redirection) will raise an exception which an attacker can observe.