COOP does not fully mitigate Cross-window Timing attacks
The browser can only enforce COOP after it receives the response. If there is computation server-side based on request (user-dependent)-data, it is still possible to measure this difference in the exact same way as in the example i.e., polling window.origin of the window handler. Instead of an exception caused by the SOP, we get an exception caused by COOP enforcement but it is detectable nonetheless.


I believe the same reasoning can be applied to timeless timing attacks.